Use UIFontMetrics to scale fonts
On iOS 11 and up, use UIFontMetrics to scale font sizes. Falls back to
a calculation using the current content size category / large as a
multiplier.

Also adds the `scalingTextStyle` attribute to `TextStyle` with a default
value of `body` so that users may specify how they'd prefer to scale the
specific font.

This should not be a breaking change for anyone since it's an added
field with a default value. All tests still pass without adding this new
value anywhere.